Fix reinstall test
