--- demo_umami_content/tests/src/Functional/ReinstallationMinimalTest.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Drupal\Tests\demo_umami_content\Functional;
-
-/**
- * @group demo_umami_content
- */
-class ReinstallationMinimalTest extends ReinstallationTestBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  protected $profile = 'minimal';
-
-}
